Reorganizing distribution of structblocks between apps. Renaming some classes. More storystream stuff.
