--- article/templates/article/objects/article_column.html
@@ -1,53 +1,0 @@
-{% load humanize %}
-{% load cache %}
-{% load wagtailcore_tags %}
-{% load wagtailimages_tags %}
-{% load video_filters %}
-{% load articletags %}
-
-<article class="o-article {{style}} {%if article.featured_media.first.video %}video-article{% endif %}" time="{{article.published_at|naturalday}}">
-    {% if article.featured_media.first or article.featured_media %}
-    {% if article.featured_media.first.image  or article.featured_media%}
-    <div class="o-article__left">
-      <div style="position:relative">
-        <a href="{% pageurl article %}" class="o-article__image">
-          {% if article.featured_media.first.image %}
-            {% image article.featured_media.first.image fill-680x476-c100 format-webp loading="lazy" %}
-          {% endif %}
-        </a>
-        {% if label %}
-          {% if article|get_label %}
-            <a href="{{article|get_section_link}}">
-              <img class="o-article__label" src="{{article|get_label}}">
-            </a>
-          {% endif %}
-        {% endif %}
-      </div>
-    </div>
-    {% elif article.featured_media.first.video %}
-    <div class="o-article__left">
-      <div class="o-article__video">
-        <iframe src="https://www.youtube.com/embed/{{ article.featured_media.first.video.url|youtube_embed_id|safe }}" title="YouTube video player" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture; web-share" allowfullscreen></iframe>
-      </div>
-    </div>
-    {% endif %}
-    {% endif %}
-    <div class="o-article__right">
-      <div class="o-article__meta">
-        {% if pubdate %}
-            <p class="o-article__published">{{ article.published_at|display_pubdate }}</p>
-        {% endif %}
-        <h3 class="o-article__headline">
-          <a href="{% pageurl article %}">{{ article.title|safe }}</a>
-        </h3>
-        {% if lede %}
-            <p class="o-article__snippet">{{article.lede|safe}}</p>
-        {% endif %}
-        {% if byline %}
-            <p class="o-article__byline">
-                <span class="o-article__author">{{ article.get_authors_split_out_visual_bylines|safe }}</span>
-            </p>
-        {% endif %}
-      </div>
-    </div>
-  </article>
