+++ article/templates/article/objects/article_top-headline.html
@@ -8,7 +8,7 @@
 
 <article class="headline-top-articles">
     <div class="above-headline"><span class="o-article__published">{{ article.published_at|display_pubdate }}</span> {% if article.title_tag_str %}<span class="title-tag">{{article.title_tag_str}}//</span>{% endif %}</div>
-    <h3><a href="{% pageurl article %}">{{article.title|safe}}</a></h3>
+    <h3 class="o-article__headline"><a href="{% pageurl article %}">{{article.title|safe}}</a></h3>
     <div class="o-article--default">
         <div class="left">
             <a href="{% pageurl article %}">
