+++ article/templates/article/objects/infinitefeed_item.html
@@ -30,18 +30,16 @@
 	{% endif %}
 	{% endif %}
 	<div class="o-article__right">
-		<div class="o-article__meta">
-			<div class="above-headline">
-				<span class="o-article__published">{{ article.published_at|display_pubdate }}</span>{%if expectedSection != article.current_section %}<a class="o-article__section" href="{{article|get_section_link}}">{{article|get_section_title}}</a>{% endif %}
-			</div>
-			<h3 class="o-article__headline">
-				<a href="{% pageurl article %}">{{ article.title|safe }}</a>
-			</h3>
-			<p class="o-article__snippet">{{ article.lede|safe }}</p>
-			<p class="o-article__byline">
-				<span class="o-article__author">{{ article.get_authors_split_out_visual_bylines|safe }}</span>
-			</p>
+		<div class="above-headline">
+			<span class="o-article__published">{{ article.published_at|display_pubdate }}</span>{%if expectedSection != article.current_section %}<a class="o-article__section" href="{{article|get_section_link}}">{{article|get_section_title}}</a>{% endif %}
 		</div>
+		<h3 class="o-article__headline">
+			<a href="{% pageurl article %}">{{ article.title|safe }}</a>
+		</h3>
+		<p class="o-article__snippet">{{ article.lede|safe }}</p>
+		<p class="o-article__byline">
+			<span class="o-article__author">{{ article.get_authors_split_out_visual_bylines|safe }}</span>
+		</p>
 	</div>
 </article>
 {% endif  %}
