+++ article/templates/article/objects/storystream_views/storystream_attachments/pdf.html
@@ -1,0 +1,1 @@
+<iframe class="pdf" src="{{self.pdf.url}}" width="800" height="500"></iframe>
