+++ events/templates/events/objects/homepage-events-bar.html
@@ -1,0 +1,16 @@
+<section class="o-events-bar">
+    <div class="o-events-bar-header">
+        <h2><a href="/events/?hidden=seminar">Events Around Campus</a></h2>
+        <p class="see_more"><a href="/events/?hidden=seminar">See our calendar for more</a></p>
+    </div>
+
+    <ul>
+        {% for event in events %}
+        <li class="{{event.category}} o-event">
+            <a href="/events/?event={{event.hash}}&category={{event.category}}" title="{{event.title}}">
+            <b>{{event.display_time}}</b> {{event.title|safe}}
+            </a>
+        </li>
+        {% endfor %}
+    </ul>
+</section>
