+++ events/templates/events/objects/sidebar_events_block.html
@@ -1,8 +1,8 @@
-<div class="events-sidebar">
+<div class="o-events-sidebar">
     <h2><a href="/events/?hidden=seminar">Events Around Campus</a></h2>
     <ul>
     {% for event in ongoing %}
-    <li class="{{event.category}}">
+    <li class="{{event.category}} event">
         <a class="featured_event" href="/events/?event={{event.hash}}&hidden=seminar" title="{{event.title}}">
         <p class="featured_event__time">{{event.display_time}}{% if event.location %}{% endif %}</p>
         <p class="featured_event__title">{{event.title|safe}}</p>
@@ -15,7 +15,7 @@
         <h3>Upcoming</h3>
         <ul>
         {% for event in upcoming %}
-            <li class="{{event.category}}">
+            <li class="{{event.category}} event">
                 <a href="/events/?event={{event.hash}}&hidden=seminar" title="{{event.title}}">
                 <b>{{event.display_time}}</b> {{event.title|safe}}
                 </a>
