+++ home/templates/home/stream_blocks/links.html
@@ -15,7 +15,7 @@
 			</ul>
 		</div>
 		<div class="home_link_ad">
-			{% include "infinitefeed/sidebar/sidebar_events_block.html" %}
+			{% include "events/objects/sidebar_events_block.html" %}
 		</div>
 	{% else %}
 		<div class="sidebar-block sidebar-block--advertisement o-advertisement--leaderboard"></div>
