--- home/templates/home/stream_blocks/section_block.html
@@ -1,13 +1,0 @@
-{% load wagtailcore_tags %}
-
-
-<section class="c-homepage__section">
-    <h1 class="section">
-        <a href={% pageurl section %}>{{ section|title }}</a>
-    </h1>
-    {% if layout == 'bulleted' %}
-    {% include "section/objects/section_bulleted.html" %}
-    {% else %}
-        {% include 'section/objects/section_featured.html' with featured_articles=articles %}
-    {% endif %}
-</section>
