+++ infinitefeed/blocks.py
@@ -3,7 +3,7 @@ from wagtail.blocks import field_block
 from wagtail.fields import StreamField
 from wagtail.images.blocks import ImageChooserBlock
 from wagtail.snippets.blocks import SnippetChooserBlock
-from article.models import ArticlePage
+from article.blocks_outer_article import ArticleGathererBlock, ManualArticles 
 from django.db.models import Q
 from django.utils.safestring import mark_safe
 from django.template.loader import render_to_string
@@ -80,60 +80,21 @@ class SidebarFlexStreamBlock(blocks.StructBlock):
         verbose_name = "Sidebar Stream Flex Block"
         verbose_name_plural = "Sidebar Stream Flex Blocks"
 
-class SidebarEventsBlock(blocks.StructBlock):
+    
+class SidebarArticleGatherer(ArticleGathererBlock):
+
+    template = SideBarListTemplates()
 
     def get_context(self, value, parent_context=None):
-        from events.models import Event
-        from django.utils import timezone
-        from datetime import timedelta
-        context = super().get_context(value, parent_context)
-        events = Event.objects.filter(hidden=False, end_time__gte=timezone.now()).exclude(category='seminar').order_by("start_time")[:15]
-        context["ongoing"] = []
-        context["upcoming"] = []
-        today = timezone.now().astimezone(timezone.get_current_timezone())
-        for i in range(len(events)):
-            
-            if events[i].start_time < today:
-                pubdate = events[i].end_time.astimezone(timezone.get_current_timezone())
-                display = "Ends "
-            else:
-                if len(context["ongoing"]) + len(context["upcoming"]) > 5:
-                    break
-                pubdate = events[i].start_time.astimezone(timezone.get_current_timezone())
-                display = ""
-                
-            delta = abs(today - pubdate)
-
-            if pubdate.date() == today.date():
-                day = ""
-            elif (pubdate - timedelta(days=1)).date() == today.date():
-                day = "Tomorrow"
-            elif delta.total_seconds() < timedelta(days=6).total_seconds():
-                if events[i].start_time < today:
-                    day = pubdate.strftime("%A")
-                else:
-                    day = pubdate.strftime("%a")
-            else:
-                day = pubdate.strftime("%B %-d") + ","
-
-            time = ""
-            if pubdate.hour != 0 and pubdate.hour != 23:
-                time = " " + pubdate.strftime("%-I")
-                if pubdate.strftime("%M") != "00":
-                    time = time + pubdate.strftime(":%M")
-                time = time + pubdate.strftime("%P")
-
-            display = display + day + time
-            events[i].display_time = display
-            events[i].title = events[i].title.replace("<br>", ", ")
-
-            if events[i].start_time < today:
-                context["ongoing"].append(events[i])
-            else:
-                context["upcoming"].append(events[i])
-        
-        context["ongoing"].sort(key=lambda e: e.end_time)
+        context = super().get_context(value, parent_context=parent_context)
+        context['articles'] = context['articles'][:5]        
         return context
 
-    class Meta:
-        template = "infinitefeed/sidebar/sidebar_events_block.html"
+class SidebarManualArticles(ManualArticles):
+
+    template = SideBarListTemplates()
+
+    def get_context(self, value, parent_context=None):
+        context = super().get_context(value, parent_context=parent_context)
+        context['articles'] = context['articles'][:5]        
+        return context
