+++ section/templates/section/objects/section_article-top-row-below.html
@@ -1,0 +1,18 @@
+{% comment %} 
+Arguments: featured_articles a list of article pages
+Used in both section pages and homepage
+{% endcomment %}
+{% load wagtailcore_tags %}
+
+<section class="landing_stream_block o-section_article-top-row-below">
+  <h2 class="section">
+      <a href={{ link }}>{{ title|title }}</a>
+  </h2>
+  <div class="o-section_article-top-row-below__row">
+    {% for article in articles|slice:"1:" %}
+      {% with pubdate="true" byline="true" style="o-article--suggested_article" %}
+      {% include 'article/objects/article_left-image.html' %}
+      {% endwith %}
+    {% endfor %}
+  </div>
+</section>
