+++ section/templates/section/objects/section_horizontal.html
@@ -16,7 +16,7 @@
   {% endif %}
 <div class="landing_stream_block--flex-container {{flex_style}}">
   {% for article in articles %}
-    {% include 'article/objects/article_column.html' with article=article style='article--round_with_border' pubdate="true" %}
+    {% include 'article/objects/article_left-image.html' with article=article style='article--round_with_border' pubdate="true" %}
   {% endfor %}
 </div>
