+++ section/templates/section/objects/section_timeline.html
@@ -14,7 +14,7 @@
   </div>
 <div class="landing_stream_block--flex-container">
   {% if articles.0 %}
-    {% include 'article/objects/article_column.html' with article=articles.0 style='article--round_with_border' pubdate="true" %}
+    {% include 'article/objects/article_left-image.html' with article=articles.0 style='article--round_with_border' pubdate="true" %}
   {% endif %}
 
 	<ul class="vertical-timeline wrapping">
