+++ section/templates/section/section_page.html
@@ -115,7 +115,7 @@
         </div>
       </div>
 
-      <div class="c-infinitefeed__feed">
+      <div class="o-feed-item">
         <div class="lightmode section_infinitefeed">
 
         {% if error %}
