+++ ubyssey/static_src/src/styles/components/events/_events-homepage-blocks.scss
@@ -1,0 +1,139 @@
+@use '../../modules/fonts';
+@use "../../modules/variables";
+
+.o-events-sidebar {
+    border: 1px solid;
+    padding: 0 1em;
+    border-radius: 0.5em;
+
+    h2, h3 {
+        font-family: fonts.$font-default;
+        font-size: 1.17em;
+    }
+    p {
+        color: inherit;
+    }
+    .featured_event {
+        &:hover * {
+            color: variables.$color-ubyssey-blue;
+        } 
+        .featured_event__time {
+            margin-block: 0;
+            font-weight: 600;
+            font-variation-settings: "wght" 600;
+            text-decoration: underline;
+        }
+        .featured_event__title {
+            margin-block: 0;
+        }
+    }
+    ul {
+        padding: 0;
+        list-style: none;
+    }
+    .see_more {
+        font-size: 0.8em;
+        text-decoration: underline;
+    }
+}
+
+li.o-event {
+    margin-bottom: 0.5em;
+    padding-left: 0.5em;
+    border-left: 0.25em solid;
+    display: -webkit-box;
+    a {
+        display: -webkit-box;
+        height: fit-content;
+        overflow: clip;
+        font-family: fonts.$font-headline;
+        font-variation-settings: "wght" 400;
+        text-overflow: ellipsis;
+        -webkit-line-clamp: 3;
+        -webkit-box-orient: vertical;
+        b {
+            font-variation-settings: "wght" 600;
+        }
+    }
+
+    &.community {
+        border-color: rgb(150,200,250);
+    }
+    &.entertainment {
+        border-color: rgb(200,250,199)
+    }
+    &.sports {
+        border-color: rgb(250,200,150)
+    }
+    &.seminar {
+        border-color: rgb(199,150,200);
+    }
+}
+
+[color-css-theme="dark"] .o-events-sidebar {
+    color: variables.$color-ubyssey-blue-light;
+    border-color: #5b7480;
+
+    h2, h3 {
+        color: white;
+    }
+}
+
+[color-css-theme="light"] .o-events-sidebar {
+    color: #3b3b3b;
+    border: var(--contrast-color) 2px solid;
+    background-color: white;
+    filter:drop-shadow(-7px 10px 4px rgba(4, 22, 32, 0.1098039216));
+
+    h2, h3 {
+        color: #222;
+    }
+}
+
+.o-events-bar {
+    margin: 2em auto;
+    padding: 0 1em;
+    display: flex;
+    max-width: 1200px;
+    border-block: 1px solid;
+    color: var(--contrast-color);
+  
+    .o-events-bar-header {
+        margin-top: 0.5em;
+        margin-right: 1em;
+        h2 {
+            margin-block: 0.5em;
+            width: fit-content;
+            font-family: fonts.$font-default;
+            font-size: 1.17em;
+            color: var(--header_color)
+        }
+        p {
+            margin-top: 0;
+            width: fit-content;
+            font-size: .8em;
+            text-decoration: underline;
+            color: var(--contrast-color);
+        }
+    }
+    ul {
+        display: contents;
+        li {
+            margin: 1em;
+            width: 200px;
+            height: auto;
+            flex-grow: 1;
+        }
+    }
+  
+    @media(variables.$bp-smaller-than-tablet) {
+        margin-block: 4em;
+        padding: 1.5em;
+        flex-direction: column;
+        ul li {
+            margin: 0.5em 0;
+            width: auto;
+        }
+    }
+  
+  }
