--- ubyssey/static_src/src/styles/components/events/_sidebar.scss
@@ -1,85 +1,0 @@
-@use '../../modules/fonts';
-@use "../../modules/variables";
-
-.events-sidebar {
-    border: 1px solid;
-    padding: 0 1em;
-    border-radius: 0.5em;
-
-    h2, h3 {
-        font-family: fonts.$font-default;
-        font-size: 1.17em;
-    }
-    p {
-        color: inherit;
-    }
-    .featured_event {
-        &:hover * {
-            color: variables.$color-ubyssey-blue;
-        } 
-        .featured_event__time {
-            margin-block: 0;
-            font-weight: 600;
-            font-variation-settings: "wght" 600;
-            text-decoration: underline;
-        }
-        .featured_event__title {
-            margin-block: 0;
-        }
-    }
-    ul {
-        padding: 0;
-        list-style: none;
-        li {
-            margin-bottom: 0.5em;
-            padding-left: 0.5em;
-            display: -webkit-box;
-            border-left: 0.25em solid;
-
-            overflow: clip;
-            font-family: fonts.$font-headline;
-            font-variation-settings: "wght" 400;
-            text-overflow: ellipsis;
-            -webkit-line-clamp: 3;
-            -webkit-box-orient: vertical;
-
-            b {
-                font-variation-settings: "wght" 600;
-            }
-
-            &.community {
-                border-color: rgb(150,200,250);
-            }
-            &.entertainment {
-                border-color: rgb(200,250,199)
-            }
-            &.sports {
-                border-color: rgb(250,200,150)
-            }
-        }
-    }
-    .see_more {
-        font-size: 0.8em;
-        text-decoration: underline;
-    }
-}
-
-[color-css-theme="dark"] .events-sidebar {
-    color: variables.$color-ubyssey-blue-light;
-    border-color: #5b7480;
-
-    h2, h3 {
-        color: white;
-    }
-}
-
-[color-css-theme="light"] .events-sidebar {
-    color: #3b3b3b;
-    border: var(--contrast-color) 2px solid;
-    background-color: white;
-    filter:drop-shadow(-7px 10px 4px rgba(4, 22, 32, 0.1098039216));
-
-    h2, h3 {
-        color: #222;
-    }
-}
