+++ ubyssey/static_src/src/styles/home.scss
@@ -35,6 +35,6 @@
 @use 'components/replacement';
 @use 'components/infinitefeed';
 
-@use 'components/events/sidebar';
+@use 'components/events/events-homepage-blocks';
 
 @use 'modules/spoof/2025/homepage-promo';
