+++ ubyssey/static_src/src/styles/objects/_box-label.scss
@@ -5,7 +5,7 @@
     top: 20px;
 }
   
-.boxLabel {
+.o-boxLabel {
     z-index: 100;
     display:flex;
     margin: auto;
@@ -14,13 +14,13 @@
       top: 75px;
     }
 }
-.boxLabel-left {
+.o-boxLabel-left {
     width: 40%;
     padding-left: 2%;
     border-width: 0.1em;
     border-style: solid;
 }
-.boxLabel-right {
+.o-boxLabel-right {
     width: 60%;
     text-align: right;
     padding-right: 2%;
@@ -28,7 +28,7 @@
     border-style: solid;
 }
   
-.boxLabel-right strong {
+.o-boxLabel-right strong {
     white-space: pre;
 }
   
@@ -40,14 +40,14 @@
     transform: translate(0,-50%);
 }
   
-.boxLabel-left {
+.o-boxLabel-left {
     background-color: var(--boxPrimary);
     border-color: var(--boxPrimary);
     .boxText-left {
         color:var(--boxTextLeft);
     }
 }
-.boxLabel-right {
+.o-boxLabel-right {
     background-color: var(--boxSecondary);
     border-color: var(--boxPrimary);
     .boxText-right {
@@ -56,18 +56,18 @@
 }  
 
 [color-css-theme="light"] {
-    .boxLabel {
+    .o-boxLabel {
         --boxPrimary: #{variables.$color-ubyssey-blue-dark};
         --boxSecondary: none;
         --boxTextLeft: white;
         --boxTextRight: #{variables.$color-ubyssey-blue-dark};
     }
 
-    .boxLabel.boxLabel-highlight {
+    .o-boxLabel.o-boxLabel-highlight {
         --boxSecondary: #{variables.$color-ubyssey-blue-light};
     }
 
-    .boxLabel {
+    .o-boxLabel {
         &.sticky {
             filter: drop-shadow(0 5px 6px #00000040);
         }
@@ -76,13 +76,13 @@
   
 [color-css-theme="dark"] {
   
-    .boxLabel {
+    .o-boxLabel {
         &.sticky {
             box-shadow: 0 5px 20px 6px variables.$color-ubyssey-blue-dark;
         }
     }
   
-    .boxLabel {
+    .o-boxLabel {
         --boxPrimary: #{variables.$color-ubyssey-blue-light};
         --boxSecondary: #{variables.$color-ubyssey-blue-dark};
         --boxTextLeft: #{variables.$color-ubyssey-blue-dark-2};
