+++ ubyssey/templates/tag/tag_page.html
@@ -38,7 +38,7 @@
 
 		<div class="u-container u-container--mid">
 	
-			<div class="c-infinitefeed__feed">
+			<div class="o-feed-item">
 			<div class="c-page__header tag-page__title">
 				<h1 class="c-page__heading u-text-center">{{meta.title}}</h1>
 				<p>
